# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Oracle API
  description: AI-powered legal intelligence platform
  version: 1.0.0
  contact:
    name: Legal Oracle Team
    email: support@legaloracle.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://legal-oracle.netlify.app
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/v1/jurisdiction/optimize:
    post:
      tags:
        - oracle
      summary: Optimize jurisdiction selection
      description: Recommend optimal jurisdiction for a case
      operationId: optimizeJurisdiction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JurisdictionRequest'
      responses:
        '200':
          description: Jurisdiction recommendation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurisdictionResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Gemini service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/compliance/optimize:
    post:
      tags:
        - oracle
      summary: Optimize compliance strategy
      description: Generate compliance optimization recommendations
      operationId: optimizeCompliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceRequest'
      responses:
        '200':
          description: Compliance recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Gemini service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/trends/forecast:
    post:
      tags:
        - oracle
      summary: Forecast regulatory changes
      description: Forecast upcoming regulatory changes
      operationId: forecastTrends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrendsRequest'
      responses:
        '200':
          description: Regulatory forecast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Gemini service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/precedent/simulate:
    post:
      tags:
        - oracle
      summary: Simulate precedent impact
      description: Simulate the impact of a legal precedent
      operationId: simulatePrecedent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecedentRequest'
      responses:
        '200':
          description: Precedent impact analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecedentResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Gemini service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/arbitrage/alerts:
    post:
      tags:
        - oracle
      summary: Generate arbitrage alerts
      description: Generate temporary legal advantages or loopholes for a user
      operationId: generateArbitrageAlerts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_role:
                  type: string
                  enum: [individual, business, lawyer, researcher]
                jurisdiction:
                  type: string
                legal_interests:
                  type: array
                  items:
                    type: string
                alert_frequency:
                  type: string
      responses:
        '200':
          description: Arbitrage opportunities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArbitrageResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Gemini service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/trends/model:
    post:
      tags:
        - oracle
      summary: Model legal evolution
      description: Model trends in legal evolution
      operationId: modelLegalEvolution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvolutionRequest'
      responses:
        '200':
          description: Legal evolution model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvolutionResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Gemini service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JurisdictionRequest:
      type: object
      required:
        - case_type
        - key_facts
      properties:
        case_type:
          type: string
        key_facts:
          type: string
    JurisdictionResponse:
      type: object
      properties:
        recommended:
          type: string
        rationale:
          type: string
    ComplianceRequest:
      type: object
      required:
        - industry
      properties:
        industry:
          type: string
        jurisdiction:
          type: string
          nullable: true
    ComplianceResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            type: string
    TrendsRequest:
      type: object
      required:
        - industry
      properties:
        industry:
          type: string
        jurisdiction:
          type: string
          nullable: true
    TrendsResponse:
      type: object
      properties:
        forecast:
          type: string
    PrecedentRequest:
      type: object
      required:
        - case_id
        - decision
        - jurisdiction
      properties:
        case_id:
          type: string
        decision:
          type: string
        jurisdiction:
          type: string
    PrecedentResponse:
      type: object
      properties:
        impact_analysis:
          type: string
        affected_cases:
          type: array
          items:
            type: string
...
    ArbitrageResponse:
      type: object
      properties:
        opportunities:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              description:
                type: string
              urgency:
                type: string
              expiration:
                type: string
                format: date-time
    EvolutionRequest:
      type: object
      required:
        - legal_domain
        - time_horizon
      properties:
        legal_domain:
          type: string
        time_horizon:
          type: string
    EvolutionResponse:
      type: object
      properties:
        trend_analysis:
          type: string
        evolution_model:
          type: object
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
          description: Optional error code for programmatic handling